+++ src/components/HistoryTables/MessageHistory/History/MessageConfirmedRow.tsx
@@ -1,17 +1,20 @@
 import React, { useMemo } from 'react'
 import { SubscriptionResult } from '@apollo/client'
 import PropTypes from 'prop-types'
-import { TR, TD } from '../table'
-import { Badge } from '../generic'
-import { SmartLink } from '../../Link/SmartLink'
-import { AddressLink } from '../../AddressLink'
-import { MessageConfirmedRow, MESSAGE_CONFIRMED_ROW_PROP_TYPE } from '../types'
-import { attoFilToFil } from '../utils'
-import { ChainHeadSubscription } from '../../../generated/graphql'
-import { useAge } from './useAge'
-import { useMethodName } from './useMethodName'
-import { isAddrEqual } from '../../../utils/isAddrEqual'
-import truncateAddress from '../../../utils/truncateAddress'
+import { TR, TD } from '../../table'
+import { Badge } from '../../generic'
+import { SmartLink } from '../../../Link/SmartLink'
+import { AddressLink } from '../../../AddressLink'
+import {
+  MessageConfirmedRow,
+  MESSAGE_CONFIRMED_ROW_PROP_TYPE
+} from '../../types'
+import { attoFilToFil } from '../../utils'
+import { ChainHeadSubscription } from '../../../../generated/graphql'
+import { useAge } from '../hooks/useAge'
+import { useMethodName } from '../hooks/useMethodName'
+import { isAddrEqual } from '../../../../utils/isAddrEqual'
+import truncateAddress from '../../../../utils/truncateAddress'
 
 export default function MessageHistoryRow(props: MessageHistoryRowProps) {
   const { message, cidHref, inspectingAddress } = props
