+++ src/components/HistoryTables/MessageHistory/History/MessagePendingRow.tsx
@@ -3,14 +3,14 @@ import { FilecoinNumber } from '@glif/filecoin-number'
 import PropTypes from 'prop-types'
 import * as dayjs from 'dayjs'
 import * as relativeTime from 'dayjs/plugin/relativeTime'
-import { Badge } from '../generic'
-import { TR, TD } from '../table'
-import { SmartLink } from '../../Link/SmartLink'
-import { AddressLink } from '../../AddressLink'
-import { MessagePendingRow, MESSAGE_PENDING_ROW_PROP_TYPE } from '../types'
-import { useMethodName } from './useMethodName'
-import { isAddrEqual } from '../../../utils/isAddrEqual'
-import truncateAddress from '../../../utils/truncateAddress'
+import { Badge } from '../../generic'
+import { TR, TD } from '../../table'
+import { SmartLink } from '../../../Link/SmartLink'
+import { AddressLink } from '../../../AddressLink'
+import { MessagePendingRow, MESSAGE_PENDING_ROW_PROP_TYPE } from '../../types'
+import { useMethodName } from '../hooks/useMethodName'
+import { isAddrEqual } from '../../../../utils/isAddrEqual'
+import truncateAddress from '../../../../utils/truncateAddress'
 
 // add RelativeTime plugin to Day.js
 dayjs.extend(relativeTime.default)
