+++ src/components/HistoryTables/MessageHistory/History/MessageRowColumnTitles.tsx
@@ -1,5 +1,5 @@
 import { bool } from 'prop-types'
-import { TR, TH } from '../table'
+import { TR, TH } from '../../table'
 
 export function MessageRowColumnTitles({
   displayIncoming
