+++ src/components/HistoryTables/detail.tsx
@@ -23,7 +23,7 @@ import {
 import { Lines, Line, AddressLine } from '../Layout'
 import { PROPOSAL_ROW_PROP_TYPE } from './types'
 import { getMethodName } from './methodName'
-import { useUnformattedDateTime } from './MessageHistory/useAge'
+import { useUnformattedDateTime } from './MessageHistory/hooks/useAge'
 import { AddressLink } from '../AddressLink'
 import { attoFilToFil, formatNumber } from './utils'
 import {
