+++ src/components/index.ts
@@ -28,7 +28,7 @@ export { default as Glyph } from './Glyph'
 export { default as Box } from './Box'
 export { default as Loading } from './LoaderGlyph'
 export { default as LoadingScreen } from './LoadingScreen'
-export { default as MessageHistoryTable } from './HistoryTables/MessageHistory'
+export { default as MessageHistoryTable } from './HistoryTables/MessageHistory/History'
 export { default as MessageDetail } from './HistoryTables/MessageHistory/Detail'
 export { default as ProposalHistoryTable } from './HistoryTables/MsigProposals'
 export { default as ProposalDetail } from './HistoryTables/MsigProposals/Proposal'
